requirements.txt
requests==2.32.3
bs4==0.0.2
tzlocal==5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
